Fixed a visual artifact on ToolbarButton's border.
Fixed a visual artifact on ToolbarButton where the insets of the
border may not have updated correctly, depending on how the
animation timing ended up resolving.
This was caused by not recomputing the border when the height of
the ToolbarButton changed, resulting in incorrect border insets.

Bug: 1053917
Change-Id: Ia15ffaf7a5d08f75b4d88bc61deb7bcc3382df2d
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2067384
Commit-Queue: Thomas Lukaszewicz <tluk@chromium.org>
Reviewed-by: Peter Kasting <pkasting@chromium.org>
Cr-Commit-Position: refs/heads/master@{#745477}