Remove isSwapPixelFormatToFixConvertFromTranslucentEnabled()
Use the generic way of checking if a feature flag is enabled in
CachedFeatureFlags.

There is a slight functional change: if the value is queried before
native is loaded, and then a different value is cached when native is
loaded, a second query will now give the same result returned in the
first.

Bug: 1012975
Change-Id: I314455b927d5f4a46aeb13de922682cc600e7056
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2079532
Reviewed-by: Peter Kotwicz <pkotwicz@chromium.org>
Commit-Queue: Henrique Nakashima <hnakashima@chromium.org>
Cr-Commit-Position: refs/heads/master@{#745487}