[UKM] Split Blink's ProxyCommit into waiting and processing
The Paint Holding feature revealed that a lot of time is spent
committing main thread document update output to the compositor
impl thread. Unfortunately, we have no breakdown of the ProxyCommit
metric into the time waiting for the compositor thread to begin
processing the commit, and the actual commit processing time.

This change records the compositor commit start time and passes it
back to the layer tree host to then feed back to the local frame
UKM recorder. UMA is also updated to record the separate wait and
commit times.

Bug: 1031702
Change-Id: Ic083a94aa267b4332e402464c2a8b9b00d75aade
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2076179
Commit-Queue: Stephen Chenney <schenney@chromium.org>
Reviewed-by: Kentaro Hara <haraken@chromium.org>
Reviewed-by: Robert Kaplow <rkaplow@chromium.org>
Reviewed-by: vmpstr <vmpstr@chromium.org>
Cr-Commit-Position: refs/heads/master@{#745515}