[Autofill Assistant] Add integration test for Overlay
The existing UI tests do not test the interaction (and actual retrieval
of element positions) from the webpage. This tests the full interaction
of a FocusElement action.

The test is somewhat flawed as it would also pass if there is no
showcast at all.

Bug: b/143942385
Change-Id: Id843f55a965dc996a00001fd02c190e8247eae4e
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2078212
Reviewed-by: Clemens Arbesser <arbesser@google.com>
Commit-Queue: Sandro Maggi <sandromaggi@google.com>
Cr-Commit-Position: refs/heads/master@{#745536}