[Autofill Assistant] Stack for element position getting
Before this change, element coordinates were relative to their window,
i.e. frame. For elements in iFrame this meant that the "cutout" was
misplaced.
This change now keeps a frame-stack while retrieving elements and
then sums up the position to get a "global" position of the element
for using in a showcast.

Bug: b/143942385
Change-Id: I8f3827002df58874d7921e969954f911630cc4ae
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2074499
Commit-Queue: Sandro Maggi <sandromaggi@google.com>
Reviewed-by: Mathias Carlen <mcarlen@chromium.org>
Reviewed-by: Stephane Zermatten <szermatt@chromium.org>
Cr-Commit-Position: refs/heads/master@{#745566}