[Enumerate Devices] Disable EnumerateDevicesHideDeviceIDs on Android.
It was noted that in the case of WebView the labels are not shown even
though the permission to access the devices were granted. After further
inspection, crrev.com/c/1964495 exposed an issue in permission checking
in the case of WebViews.

To give us time to look into the issue further and fix the actual
permission checking for WebViews, this CL disables by default the behavior
change of enumerateDevices on Android.

Bug: 1055770
Change-Id: I9dc557fed9961cf90794c87c4da4fa12213793f1
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2078547
Commit-Queue: Guido Urdaneta <guidou@chromium.org>
Reviewed-by: Avi Drissman <avi@chromium.org>
Reviewed-by: Guido Urdaneta <guidou@chromium.org>
Cr-Commit-Position: refs/heads/master@{#745581}