Add Total Blocking Time to Time To Interactive's trace event
This CL adds TBT to TTI's trace event by keeping track of the longtasks
and using the list to compute TBT as soon as TTI is known. The
computation is done inside the trace event to avoid running the code
when tracing is disabled. An args param is added because there are no
macros supporting more than 2 args, so two of them are merged into a
single TracedValue param.

Bug: 1055806
Change-Id: Ided9c2378aff6d98cbdf5e88e99d46735b92e284
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2079616
Reviewed-by: Steve Kobes <skobes@chromium.org>
Reviewed-by: Deep Roy <dproy@chromium.org>
Commit-Queue: Nicolás Peña Moreno <npm@chromium.org>
Cr-Commit-Position: refs/heads/master@{#745610}