Clamp video.rAF timing information
This CL decreases the resolution of the presentationTime,
expectedPresentationTime, elapsedProcessingTime and captureTime fields,
to limit potential timing attacks.

expectedPresentationTime should correspond to an animation frame
timestamp and technically shouldn't need be clamped. However, this is
not always the case in the WebRTC path, and we should keep it clamped
for now.

Bug: 1012063
Change-Id: Ie9247be8f3c91e2fef3d62cf939b1d8488f0a75a
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2071113
Reviewed-by: Mounir Lamouri <mlamouri@chromium.org>
Reviewed-by: Daniel Vogelheim <vogelheim@chromium.org>
Commit-Queue: Thomas Guilbert <tguilbert@chromium.org>
Cr-Commit-Position: refs/heads/master@{#745612}