weblayer: converts Browser.getTabs() from returning a List to a Set
A list implies a certain order. getTabs() does not return a particular
order. This converts the return value to a Set to reinforce the order is
not defined.

This is an incomptible API change. The change is purely in the client
library, so there is worry about compatibility with different implementation
versions. This change was requested by Alihan, and they understand it is
incomptible.

BUG=none
TEST=none

Change-Id: I5c4777b6624ba501c0708eacc7d81217a17f253e
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2080478
Reviewed-by: John Abd-El-Malek <jam@chromium.org>
Commit-Queue: Scott Violet <sky@chromium.org>
Cr-Commit-Position: refs/heads/master@{#745643}