[fuchsia] Allow the process' ComponentContext to be overridden by tests.
ComponentContextForCurrentProcess() returns a sys::ComponentContext
connected to the process' incoming service directory, outgoing directory
for services & debugging APIs, etc.

Provide a TestComponentContextForProcess, which overrides the process'
ComponentContext with one which by default has no services available,
and which allows services to be provided by the test itself, and for
services available in the test process environment to be exposed to the
test.

This is a re-land of
https://chromium-review.googlesource.com/c/chromium/src/+/2064908
with fixed Debug build.
Also added TODO to migrate to ComponentContextProvider.

Bug: 1038786
Change-Id: I401ce50a948b5426ac3b4ebb6e9a75674edf7011
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2077912
Reviewed-by: Sergey Ulanov <sergeyu@chromium.org>
Commit-Queue: Sergey Ulanov <sergeyu@chromium.org>
Auto-Submit: Wez <wez@chromium.org>
Cr-Commit-Position: refs/heads/master@{#745664}