Enable swipe from login shelf for marketing opt in screen
Sets up login shelf gesture handler when marketing opt in screen is
shown, which enables swipe gesture from the login shelf (and shows a
shelf drag handle). Swipe handler will exit the screen.

The handler is set up only if shelf navigation buttons would be hidden
in the current environment - if tablet mode is on, and the buttons are
not shown for accessibility. When the swipe is enabled, the screen's
next button is hidden.

While here, add LoginScreenBehavior to marketing opt in screen polymer
element.

BUG=976949

Change-Id: Idcc2828f21c85162be29c5e4706ea3281f1b6bbc
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2080008
Reviewed-by: Alexander Alekseev <alemate@chromium.org>
Commit-Queue: Toni Baržić <tbarzic@chromium.org>
Cr-Commit-Position: refs/heads/master@{#745715}