Fix SetSize & PosInset for items nested in list.
Consider the following web content where an ordered set item is nested
in a list item:
<ul role="tablist">
  <li>
    <div role="tab" aria-posinset="3" aria-setsize="5">
      item 3 of 5
    </div>
  </li>
  <li>
    <div role="tab" aria-posinset="4" aria-setsize="5">
      item 4 of 5
    </div>
  </li>
  <li>
    <div role="tab" aria-posinset="5" aria-setsize="5">
      item 5 of 5
    </div>
  </li>
</ul>

We currently would get PosInSet=0 and SetSize=0 for the tab item in
accessibility tree.
The reason we are not getting PosInSet and SetSize value is because
AXTree::RecursivelyPopulateOrderedSetItems is not recursing pass the
list item element to add tree item to the ordered set for calculating
PosInSet and SetSize. This change fixes this issue by allowing
AXTree::RecursivelyPopulateOrderedSetItems to recurse through these
parent elements and adding set items to ordered set, thereby giving
correct calculation for PosInSet and SetSize.

Bug: 1047025
Change-Id: Ieaf31db1e0b70c84923e2b10e5ba4413ba413696
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2029189
Commit-Queue: Victor Fei <vicfei@microsoft.com>
Reviewed-by: Aaron Leventhal <aleventhal@chromium.org>
Reviewed-by: Ian Prest <iapres@microsoft.com>
Reviewed-by: Akihiro Ota <akihiroota@chromium.org>
Cr-Commit-Position: refs/heads/master@{#745716}