content: fix DCHECK in BackgroundFetchDelegateProxy
If there is no delegate GetPermissionForOrigin() immediately runs the
supplied callback. This triggers a DCHECK in BackgroundFetchContext as
BackgroundFetchContext assumes it is called on the service-worker
thread.

This is triggering a crash in wpt tests on WebLayer as we haven't wired
up background support.

BUG=1049426
TEST=covered by wpt

Change-Id: Ib72727cfa6dfeae0aa934eb03c0de533e07ccddf
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2080864
Commit-Queue: Peter Beverloo <peter@chromium.org>
Auto-Submit: Scott Violet <sky@chromium.org>
Reviewed-by: Peter Beverloo <peter@chromium.org>
Cr-Commit-Position: refs/heads/master@{#745806}