[Reader Mode] Uses showModal for settings dialog
The desired UI has been updated to center the settings dialog on the
page, and be vertically aligned with the toggle. This means we can make
use of showModal() over show() and use the native keyboard handling.

The font size slider has also been updated to dynamically show font size
changes while the mouse is being held down, so a user can scrub the
slider and immediately see changes in the font size.

Bug: 1015200, 1056964
Change-Id: I72319b7b9092af360697eedb7fb6a5f3c36d26c1
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2072343
Reviewed-by: Wei-Yin Chen (陳威尹) <wychen@chromium.org>
Reviewed-by: Alice Boxhall <aboxhall@chromium.org>
Commit-Queue: Meredith Lane <meredithl@chromium.org>
Cr-Commit-Position: refs/heads/master@{#745867}