Fix settings mismatch for new child users enabling extension installs
"Permissions for sites, apps and extensions" is a toggle setting in
families.google.com/manage/family that enables supervised users to ask
for approval to install extensions when enabled, and blocks extension
installs otherwise, among other side effects. For new child accounts,
the default setting for this toggle is on. This CL changes the default
value of the kSupervisedUserExtensionsMayRequestPermissions pref to
true on the Chrome side to match the server default, and avoid
inconsistencies for new child accounts.

Bug: 1051739
Change-Id: I80fdd4fba440ea1f6593161cdb2ef1938643a55f
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2056569
Reviewed-by: Aga Wronska <agawronska@chromium.org>
Commit-Queue: Toby Huang <tobyhuang@chromium.org>
Cr-Commit-Position: refs/heads/master@{#745875}