[Android] Move UrlUtilitiesUnitTest to be on //components-level class
A previous CL copied UrlUtilities.java into //components in preparation
for deleting the //chrome-level class once the internal repo had also
been converted. This CL moves the unittest to be on the
//components-level class; a followup will delete the //chrome-level
class. (The move of the test is involved enough that I wanted to focus
this CL just on that).

Bug: 1031465
Change-Id: I43aaebb21f7777bbc4cb8c50a3791abd793f0a8f
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2080423
Reviewed-by: Bo <boliu@chromium.org>
Commit-Queue: Colin Blundell <blundell@chromium.org>
Cr-Commit-Position: refs/heads/master@{#745879}