Report COEP errors for resource loads and navigations
This CL adds specific error reason reporting for resources that get
blocked due to COEP errors. Blocking either occurs in a URLLoader
or in a NavigationRequest. The error is plumbed to the DevTools
front-end, which should help developers identify and fix COEP issues.

Note that this CL does not address CORB blocked fetches, so DevTools
will report a plain failure if those get blocked.

Bug: chromium:1051473
Change-Id: Idad0d1669d0f45a7116419e1935feb0bb9973e91
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2070409
Reviewed-by: Camille Lamy <clamy@chromium.org>
Reviewed-by: Łukasz Anforowicz <lukasza@chromium.org>
Reviewed-by: Kinuko Yasuda <kinuko@chromium.org>
Reviewed-by: Mike West <mkwst@chromium.org>
Commit-Queue: Sigurd Schneider <sigurds@chromium.org>
Cr-Commit-Position: refs/heads/master@{#745916}