Add a common fixture for Trust Tokens operations
There's a bunch of boilerplate common to testing the three Trust Tokens
operations of issuance, redemption, and signing, namely URLRequest
creation (which involves constructing a TestURLRequestContext and a
couple includes for RequestPriority and NETWORK_ANNOTATION_FOR_TESTS)
and waiting for an asynchronous
operation---TrustTokenRequestHelper::Begin---to conclude. This CL
factors out the boilerplate into a fixture intended to be common to the
three operations' tests.

R=csharrison

Bug: 1042962
Change-Id: Id513991f38a21c1cbb64924481f38f805fb88124
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2070824
Commit-Queue: David Van Cleve <davidvc@chromium.org>
Reviewed-by: Charlie Harrison <csharrison@chromium.org>
Cr-Commit-Position: refs/heads/master@{#745951}