Add test for replacing a composited animation
CL 1715764 introduced a performance regression, where a replaced
animation would not be composited for the initial frame.  This was
caused by a bug in calculation of the pending state, which in turn
caused a replaced animation to be flagged as incompatible.  The bug was
fixed in CL 1728074. This patch introduced a test to ensure that the fix
does not regress.

Bug: 992408
Change-Id: Iab210ca9b525773ebb2c0a6f28d63f25c0b574e9
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2080445
Reviewed-by: Majid Valipour <majidvp@chromium.org>
Commit-Queue: Kevin Ellis <kevers@chromium.org>
Cr-Commit-Position: refs/heads/master@{#745970}