[Security] Removing window.name after a COOP switch.
This patch clears a frame name when a BrowsingInstance swap
happened because of COOP. This happens as follows:
1. All COOP navigations force a browsing instance swap.
2. When swapping browsing instances, a new provisional frame needs to
be created.
3. When creating a new provisional frame in response to a COOP-forced swap, the
   frame is created with a FrameReplicationState with an empty name/unique_name.
4. When the provisional frame is committed, this is detected in CommitPending(),
   which then clears the name/unique_name on the FrameTreeNode and replicates
   the name updates.

- Fixes an issue in blink where the provisional frame would use the
previous name instead of the replicated_state one passed as an argument.
- Remembers when we had a switch that was COOP related and stores this
state in the NavigationRequest.
- When we create a RenderFrame and NavigationRequest has this state, we
pass an empty name to the RenderFrame creation in the renderer.
- When we get a DidCommit from the renderer and NavigationRequest has
this state we update the name on the browser side in the FrameTreeNode.




Together this should ensure that the new RenderFrame, as well as all the
proxy references to it have the correct empty name at the end of the
navigation.

Bug: 1018655
Change-Id: I9262ba49d3c883e55a29aba1ca0d4e2011a1289c
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2027456
Commit-Queue: Arthur Hemery <ahemery@chromium.org>
Reviewed-by: Alex Moshchuk <alexmos@chromium.org>
Reviewed-by: Daniel Cheng <dcheng@chromium.org>
Reviewed-by: Pâris Meuleman <pmeuleman@chromium.org>
Cr-Commit-Position: refs/heads/master@{#745990}