Hide zoom toolbar until the UI text direction is set.
If the browser is using an RTL language, the zoom toolbar should be
displayed on the bottom left of the PDF viewer. As an extension, the
PDF viewer gets its UI directions after loading translated strings from
the backend with a latency. Due to this latency, the zoom toolbar will
first show up on the right side of the PDF viewer by default and then
switch to the left side after it gets all UI directions from the
backend.

To fix this issue, this CL will hide the zoom toolbar until all
translated strings are loaded from the backend.

Bug: 1038022
Change-Id: I6f49b888466fe9f4717c89c7b996feeda2777380
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2078747
Reviewed-by: Rebekah Potter <rbpotter@chromium.org>
Commit-Queue: Hui Yingst <nigi@chromium.org>
Cr-Commit-Position: refs/heads/master@{#746001}