[Android] Create tests for SyncErrorInfoBar
Add tests to check whether the infobar is shown for different errors. This
cl also registers SyncErrorInfoBar as a ProfileSyncService state change
observer to ensure that the infobar is dismissed as soon as the user resolves
the issue from settings.

Render tests will be added in a subsequent cl.

Bug: 1031632
Change-Id: Ifb6b64e187ddce6129a43b0a7be1d76d8ff26ace
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2033244
Reviewed-by: Boris Sazonov <bsazonov@chromium.org>
Commit-Queue: Tanmoy Mollik <triploblastic@chromium.org>
Cr-Commit-Position: refs/heads/master@{#746036}