Componentize FileTypePolicies
This CL moves the FileTypePolicies and associated code to
components/safe_browsing/core. This is useful for fixing our measurement
of the download warning CTR. In a future CL, I will refer to this code
from components/download to record click throughs based on the UMA
value for the download.

Bug: 1053772
Change-Id: I47708b0395eee9997ebdaf01435993107e23b687
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2076416
Reviewed-by: Nico Weber <thakis@chromium.org>
Reviewed-by: Ted Choc <tedchoc@chromium.org>
Reviewed-by: Bettina Dea <bdea@chromium.org>
Reviewed-by: Ken Rockot <rockot@google.com>
Reviewed-by: Shakti Sahu <shaktisahu@chromium.org>
Reviewed-by: Joshua Pawlicki <waffles@chromium.org>
Reviewed-by: Tommy Li <tommycli@chromium.org>
Commit-Queue: Daniel Rubery <drubery@chromium.org>
Cr-Commit-Position: refs/heads/master@{#746067}