weblayer: fix crash in Tab
ComposedBrowserControlsVisibilityDelegate (by way of ObservableSupplierImpl)
may post a task to notify the observer (the Tab in this case). If the Tab is
destroyed before the task runs, then when the task runs we get a crash as
the Tab has been destroyed (specifically mNativeTab is 0, which means the jni
call DCHECKs).

I would like a test for this, but it proved slightly tricky.

BUG=1057657
TEST=none

Change-Id: I85534c464fe11121fb9f65710fc3fb69b48b0d68
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2083813
Reviewed-by: Evan Stade <estade@chromium.org>
Commit-Queue: Scott Violet <sky@chromium.org>
Cr-Commit-Position: refs/heads/master@{#746113}