Specify which window is creating exceptions in portals WPTs
promise_rejects_dom and promise_rejects_js are strict about which
constructor is being used for the exception. In portals WPTs, we
frequently open a new window to run the test as we can't destroy the
context the test harness is running in. We now specify which window is
producing the expected errors.

Bug: 1057783
Change-Id: Ia3b4b55a0a12f0df96fc4bed45cd7c82a97ed5d2
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2083826
Reviewed-by: Adithya Srinivasan <adithyas@chromium.org>
Commit-Queue: Kevin McNee <mcnee@chromium.org>
Cr-Commit-Position: refs/heads/master@{#746124}